spring-boot-starter. Fixes #192
