--- handler-default/src/main/java/com/expediagroup/streamplatform/streamregistry/handler/DefaultHandler.java
@@ -1,32 +1,0 @@
-/**
- * Copyright (C) 2018-2019 Expedia, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.expediagroup.streamplatform.streamregistry.handler;
-
-import java.util.Optional;
-
-import com.expediagroup.streamplatform.streamregistry.model.Entity;
-
-public class DefaultHandler<E extends Entity<?>> implements Handler<E> {
-  @Override
-  public String type() {
-    return "default";
-  }
-
-  @Override
-  public E handle(E entity, Optional<? extends E> existing) {
-    return entity;
-  }
-}
