--- handler-default/src/main/java/com/expediagroup/streamplatform/streamregistry/handler/DefaultHandlerConfiguration.java
@@ -1,41 +1,0 @@
-/**
- * Copyright (C) 2018-2019 Expedia, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.expediagroup.streamplatform.streamregistry.handler;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import com.expediagroup.streamplatform.streamregistry.model.Domain;
-import com.expediagroup.streamplatform.streamregistry.model.Schema;
-import com.expediagroup.streamplatform.streamregistry.model.Stream;
-
-@Configuration
-public class DefaultHandlerConfiguration {
-  @Bean
-  Handler<Domain> defaultDomainHandler() {
-    return new DefaultHandler<>();
-  }
-
-  @Bean
-  Handler<Schema> defaultSchemaHandler() {
-    return new DefaultHandler<>();
-  }
-
-  @Bean
-  Handler<Stream> defaultStreamHandler() {
-    return new DefaultHandler<>();
-  }
-}
