+++ pom.xml
@@ -34,13 +34,11 @@
   </licenses>
 
   <modules>
-    <module>app</module>
     <module>core</module>
     <module>graphql-api</module>
     <module>graphql-client</module>
+    <module>graphql-client-reactor</module>
     <module>graphql-maven-plugin</module>
-    <module>handler-default</module>
-    <module>handler-egsp</module>
     <module>handler-spi</module>
     <module>model</module>
     <module>repository-api</module>
@@ -48,7 +46,7 @@
     <module>repository-kafka</module>
     <module>it</module>
     <module>service-api</module>
-    <module>graphql-client-reactor</module>
+    <module>spring-boot-starter</module>
   </modules>
 
   <properties>
