+++ spring-boot-starter/pom.xml
@@ -8,23 +8,28 @@
     <version>0.7.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>stream-registry-handler-default</artifactId>
+  <artifactId>stream-registry-spring-boot-starter</artifactId>
 
   <dependencies>
     <dependency>
       <groupId>com.expediagroup.streamplatform</groupId>
-      <artifactId>stream-registry-model</artifactId>
+      <artifactId>stream-registry-core</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
       <groupId>com.expediagroup.streamplatform</groupId>
-      <artifactId>stream-registry-handler-spi</artifactId>
+      <artifactId>stream-registry-graphql-api</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>com.expediagroup.streamplatform</groupId>
+      <artifactId>stream-registry-repository-kafka</artifactId>
       <version>${project.version}</version>
     </dependency>
 
     <dependency>
-      <groupId>org.springframework</groupId>
-      <artifactId>spring-context</artifactId>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter</artifactId>
     </dependency>
   </dependencies>
