+++ spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+# Auto Configure
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.expediagroup.streamplatform.streamregistry.StreamRegistryConfiguration
